% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcnmap.R
\name{cnmap}
\alias{cnmap}
\title{draw china Map with province}
\usage{
cnmap(name = NA, color = NA, title = "China Map", fill = NA)
}
\arguments{
\item{name}{value}

\item{color}{value}

\item{title}{value}

\item{fill}{value}
}
\value{
returndes
}
\description{
description
}
\examples{
  cnmap()
  cnmap(name=c("shanghai","hubei"),color=c("red","blue"))
}

