% file MASS/man/cnmapdata.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{cnmapdata}
\alias{cnmapdata}
\title{
  The Shape Data for Map of China
}
\description{
  The Shape Data for Map of China.
}
\usage{
cnmapdata
}
\source{
  National Geomatics Center of China
}
\references{
}
\keyword{datasets}
